extract smart-pilot system
