+++ modules/model/src/ship/index.ts
@@ -1,6 +1,8 @@
 export * from './armor';
 export * from './ship-state';
 export * from './chain-gun';
+export * from './radar';
+export * from './smart-pilot';
 export * from './ship-manager';
 export * from './ship-direction';
 export * from './ship-areas';
